+++ junos/resource_security_global_policy.go
@@ -485,9 +485,9 @@ func readSecurityGlobalPolicy(clt *Client, junSess *junosSession) (globalPolicyO
 						itemTrimPolicy, "match source-end-user-profile "), "\"")
 				case strings.HasPrefix(itemTrimPolicy, "then "):
 					switch {
-					case strings.HasSuffix(itemTrimPolicy, permitW),
-						strings.HasSuffix(itemTrimPolicy, "deny"),
-						strings.HasSuffix(itemTrimPolicy, "reject"):
+					case itemTrimPolicy == "then permit",
+						itemTrimPolicy == "then deny",
+						itemTrimPolicy == "then reject":
 						policy["then"] = strings.TrimPrefix(itemTrimPolicy, "then ")
 					case itemTrimPolicy == "then count":
 						policy["count"] = true