+++ CHANGELOG.md
@@ -7,6 +7,8 @@ ENHANCEMENTS:
 
 BUG FIXES:
 
+* resource/`junos_security_global_policy`: fix error reading config when an element of `permit_application_services` have the suffix `permit`, `deny` or `reject` (Fixes #430)
+
 ## 1.31.0 (October 12, 2022)
 
 FEATURES: